# shared/api/v1/openapi.yaml
openapi: 3.0.3
info:
  title: anti-yt api
  description: anti-yt api
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Localhost server
...
security:
  - CookieJwtAuth: [ ]
paths:
...
  /api/v1/auth/google:
    parameters:
      - $ref: '#/components/parameters/HeaderXRealIP'
      - $ref: '#/components/parameters/HeaderCfIpCountry'
      - $ref: '#/components/parameters/HeaderCfRay'
      - $ref: '#/components/parameters/HeaderUserAgent'
      - $ref: '#/components/parameters/HeaderDeviceFingerprint'
    get:
      summary: Google authentication
      description: Google OIDCでログインor登録
      operationId: getAuthGoogle
      security: [ ]
      tags: [ auth ]
      responses:
        "302":
          description: 302 redirect
          headers:
            Location:
              schema:
                type: string
                description: Googleの認証画面へリダイレクト
            Set-Cookie:
              schema:
                type: array
                items:
                  type: string
                description: stateとcode_verifierを指定
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
...
  /api/v1/auth/google/callback:
    parameters:
      - $ref: '#/components/parameters/HeaderXRealIP'
      - $ref: '#/components/parameters/HeaderCfIpCountry'
      - $ref: '#/components/parameters/HeaderCfRay'
      - $ref: '#/components/parameters/HeaderUserAgent'
      - $ref: '#/components/parameters/HeaderDeviceFingerprint'
    get:
      tags: [ auth ]
      summary: Google Authorization Code Callback
      description: Googleの認可コードを受け取るため
      operationId: getAuthGoogleCallback
      security: [ ]
      parameters:
        - name: code
          description: Googleから発行された認可コード
          in: query
          required: true
          schema:
            type: string
        - name: state
          in: query
          required: true
          description: CSRF対策用
          schema:
            type: string
        - name: csrf
          in: cookie
          required: true
          description: CSRF対策
          schema:
            type: string
        - name: error
          in: query
          required: false
          description: エラー
          schema:
            type: string
      responses:
        "302":
          description: 302 redirect
          headers:
            Location:
              schema:
                type: string
                description: 登録画面かダッシュボード
            Set-Cookie:
              schema:
                type: array
                items:
                  type: string
                description: アクセストークンとリフレッシュトークンとCSRF Tokenを設定。
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "413":
          $ref: "#/components/responses/PayloadTooLarge"
        "429":
          $ref: "#/components/responses/TooManyRequests"
        "500":
          $ref: "#/components/responses/InternalServerError"
components:
  responses:
    InternalServerError:
      description: 500 Internal Server Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetailError"
    NotFound:
      description: 404 Not Found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetailError"
    Unauthorized:
      description: 401 Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetailError"
    Forbidden:
      description: 403 Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetailError"
    BadRequest:
      description: 400 Bad Request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetailError"
    PayloadTooLarge:
      description: 413 Payload Too Large
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetailError"
    TooManyRequests:
      description: 429 Too Many Requests
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetailError"
  schemas:
    ProblemDetailError:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
      required: [ title, detail ]
...
  parameters:
    HeaderXRealIP:
      name: X-Real-IP
      in: header
      description: クライアントの実際のIPアドレス（プロキシ経由の場合）
      required: true
      schema:
        type: string
        format: ipv4
        example: "203.0.113.195"
    HeaderCfIpCountry:
      name: Cf-Ipcountry
      in: header
      description: Cloudflareが検出したクライアントの国コード（ISO 3166-1 alpha-2）
      required: true
      schema:
        type: string
        format: iso-3166-1-alpha-2
        minLength: 2
        maxLength: 2
    HeaderCfRay:
      name: Cf-Ray
      in: header
      description: Cloudflareリクエストの一意識別子（トレーシング用）
      required: true
      schema:
        type: string
    HeaderUserAgent:
      name: User-Agent
      in: header
      description: クライアントのブラウザ・アプリケーション情報
      required: true
      schema:
        type: string
...
    HeaderDeviceFingerprint:
      name: X-Device-Fingerprint
      in: header
      required: true
      schema:
        type: string
  securitySchemes:
    CookieJwtAuth:
      type: apiKey
      in: cookie
      name: access_token
      description: JWT Access Token in Cookie